added a check if the file is already installed to install.sh

+++ install.sh
@@ -1,8 +1,8 @@
 for f in `ls`
 do
-	if [ $f != "install.sh" ]
+	if  [ $f != "install.sh" -a ! -e $HOME/.$f ]
 	then
-		echo "installing \"$f\" to \"$HOME/.$f\"... "
-		ln -s `pwd`/$f $HOME/.$f
+	    echo "installing \"$f\" to \"$HOME/.$f\"... "
+	    ln -s `pwd`/$f $HOME/.$f
 	fi
 done
